Move migration utils to migration module
